Change const type

+++ src/syscall/syscall.go
@@ -2,16 +2,16 @@ package syscall
 
 import "unsafe"
 
-const SYS_READ int = 0
-const SYS_WRITE int = 1
-const SYS_OPEN int = 2
-const SYS_GETDENTS64 int = 217
+const SYS_READ uintptr = 0
+const SYS_WRITE uintptr = 1
+const SYS_OPEN uintptr = 2
+const SYS_GETDENTS64 uintptr = 217
 
 func Read(fd int, buf []byte) uintptr {
 	p := &buf[0]
 	_cap := cap(buf)
 	var ret uintptr
-	ret  = Syscall(uintptr(SYS_READ), uintptr(fd), uintptr(unsafe.Pointer(p)), uintptr(_cap))
+	ret  = Syscall(SYS_READ, uintptr(fd), uintptr(unsafe.Pointer(p)), uintptr(_cap))
 	return ret
 }
 
@@ -20,7 +20,7 @@ func Open(path string, mode int, perm int) uintptr {
 	buf = append(buf, 0) // add null terminator
 	p := &buf[0]
 	var ret uintptr
-	ret = Syscall(uintptr(SYS_OPEN), uintptr(unsafe.Pointer(p)), uintptr(mode), uintptr(perm))
+	ret = Syscall(SYS_OPEN, uintptr(unsafe.Pointer(p)), uintptr(mode), uintptr(perm))
 	return ret
 }
 
@@ -28,14 +28,14 @@ func Write(fd int, buf []byte) uintptr {
 	p := &buf[0]
 	_len := len(buf)
 	var ret uintptr
-	ret  = Syscall(uintptr(SYS_WRITE), uintptr(fd), uintptr(unsafe.Pointer(p)), uintptr(_len))
+	ret  = Syscall(SYS_WRITE, uintptr(fd), uintptr(unsafe.Pointer(p)), uintptr(_len))
 	return ret
 }
 
 func Getdents(fd int, buf []byte) int {
 	var _p0 unsafe.Pointer
 	_p0 = unsafe.Pointer(&buf[0])
-	nread := Syscall(uintptr(SYS_GETDENTS64), uintptr(fd), uintptr(_p0), uintptr(len(buf)))
+	nread := Syscall(SYS_GETDENTS64, uintptr(fd), uintptr(_p0), uintptr(len(buf)))
 	return int(nread)
 }
